Process: WIP
git-svn-id: https://bobcat.svn.sourceforge.net/svnroot/bobcat/trunk@316 95c995b8-3b1c-0410-acf1-adacd5d40d98

+++ bobcat/process/discontinue.cc
@@ -2,14 +2,14 @@
 
 pid_t Process::discontinue(RetPid &proc)
 {
-//    if (proc.pid > 0)
-//    {
-////std::cerr << "Discontinue " << proc.pid << endl;
+    if (proc.pid > 0)
+    {
+//std::cerr << "Discontinue " << proc.pid << endl;
 
         kill(proc.pid, SIGHUP) && kill(proc.pid, SIGHUP) && 
                                                     kill(proc.pid, SIGTERM);
         proc.pid = waitpid(proc.pid, &proc.ret, 0);
-//    }
+    }
 
     return proc.pid = 0;
 }
